Refactor: Moved 'section' modules to 'sections' dir :unamused:
Refs: BLAIS5-3651
