+++ src/components/questionnaireDetailsPage/sections/questionnaireSettingsSection.test.tsx
@@ -2,12 +2,12 @@
  * @jest-environment jsdom
  */
 
-import flushPromises from "../../tests/utils";
+import flushPromises from "../../../tests/utils";
 import { render, screen, waitFor } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import React from "react";
 import QuestionnaireSettingsSection from "./questionnaireSettingsSection";
-import { opnQuestionnaire } from "../../features/step_definitions/helpers/apiMockObjects";
+import { opnQuestionnaire } from "../../../features/step_definitions/helpers/apiMockObjects";
 import { createMemoryHistory } from "history";
 import { Router } from "react-router";
 import axios from "axios";
